pretenders/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.4
Flask-Testing==0.7.1
PyJWT==1.6.4
Flask-Bcrypt==0.7.1
flask-cors==3.0.6